Refactors connector execution
